+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ### Added
 
 - `demethodize`, `binary` function.
+- Placeholder value (`_`) to preserve place for argument.
+
+### Changed
+
+- `curry` function can now preserve places for arguments.
+- `Curry` type was rewritten. Thanks to [that great article from Pierre-Antoine Mills](https://www.freecodecamp.org/news/typescript-curry-ramda-types-f747e99744ab/).
 
 ## [0.23.0] - 2021-06-27